+++ pkg/shell/clamav.go
@@ -1,50 +1,60 @@
 package shell
 
-import (
-	"io"
-)
+import "io"
 
-type clamavCmd struct {
-	InitCmd      func() *Executable
+type clamScan struct {
+	InitCmd func() *Executable
 }
 
-// Version outputs the version of the ClamAV CLI
-// shell: `clamscan --version`
-func (s *clamavCmd) Version() *Command {
-	return NewCommand(s.InitCmd().WithArgs("--version"))
-}
-
-// Run runs a FreshClam command to update the CVD database
-//
-// shell: `freshclam`
-func (s *clamavCmd) Run() *Command {
-	exe := s.InitCmd()
+// Version of the clamscan CLI
+func (c *clamScan) Version() *Command {
+	exe := c.InitCmd().WithArgs("--version")
 	return NewCommand(exe)
 }
 
-// ScanFile runs ClamAV Scan
-//
-// shell: `clamscan -irv --scan-archive=yes --max-filesize=4000M --max-scansize=4000M --stdout ${TARGET_DIR}`
-func (s *clamavCmd) Scan(directory string) *Command {
-	args := []string{"-irv", "--scan-archive=yes", "--max-filesize=2000M", "--max-scansize=2000M", "--stdout", directory}
-	exe := s.InitCmd().WithArgs(args...)
+func (c *clamScan) Scan(targetDirectory string) *Command {
+	exe := c.InitCmd().WithArgs(
+		"--infected",
+		"--recursive",
+		"--verbose",
+		"--scan-archive=yes",
+		"--max-filesize=2000M",
+		"--max-scansize=2000M",
+		"--stdout",
+		targetDirectory,
+	)
+
 	return NewCommand(exe)
 }
 
-// FreshClam Command with custom stdout and stderr
-func FreshClamCommand(stdin io.Reader, stdout io.Writer, stderr io.Writer) *clamavCmd {
-	return &clamavCmd{
+type freshClam struct {
+	InitCmd func() *Executable
+}
+
+func (c *freshClam) Version() *Command {
+	exe := c.InitCmd().WithArgs("--version")
+	return NewCommand(exe)
+}
+
+func (c *freshClam) FreshClam() *Command {
+	exe := c.InitCmd().WithArgs("--verbose")
+	return NewCommand(exe)
+}
+
+// ClamScanCommand scans files for virus by cross-referencing them to a known virus database
+func ClamScanCommand(stdin io.Reader, stdout io.Writer, stderr io.Writer) *clamScan {
+	return &clamScan{
 		InitCmd: func() *Executable {
-			return NewExecutable("freshclam").WithIO(stdin, stdout, stderr)
+			return NewExecutable("clamscan").WithIO(stdin, stdout, stderr)
 		},
 	}
 }
 
-// ClamScan Command with custom stdout and stderr
-func ClamScanCommand(stdin io.Reader, stdout io.Writer, stderr io.Writer) *clamavCmd {
-	return &clamavCmd{
+// FreshClam updates the virus definition database
+func FreshClamCommand(stdin io.Reader, stdout io.Writer, stderr io.Writer) *freshClam {
+	return &freshClam{
 		InitCmd: func() *Executable {
-			return NewExecutable("clamscan").WithIO(stdin, stdout, stderr)
+			return NewExecutable("freshclam").WithIO(stdin, stdout, stderr)
 		},
 	}
 }