+++ pkg/shell/oras.go
@@ -14,6 +14,8 @@ func (o *orasCmd) Version() *Command {
 	return NewCommand(exe)
 }
 
+// PushBundle use the push command to push a gatecheck bundle
+// shell: `oras push --disable-path-validation --artifact-type ${BUNDLE_ARTIFACT_TYPE} ${SAST_ARTIFACTS_IMAGE} ${GATECHECK_BUNDLE} | tee log.txt`
 func (o *orasCmd) PushBundle(artifactImage string, bundleFilename string) *Command {
 	exe := o.InitCmd().WithArgs(
 		"push",
@@ -26,8 +28,6 @@ func (o *orasCmd) PushBundle(artifactImage string, bundleFilename string) *Comma
 
 	return NewCommand(exe)
 }
-
-// shell: `oras push --disable-path-validation --artifact-type ${BUNDLE_ARTIFACT_TYPE} ${SAST_ARTIFACTS_IMAGE} ${GATECHECK_BUNDLE} | tee log.txt`
 
 func OrasCommand(stdin io.Reader, stdout io.Writer, stderr io.Writer) *orasCmd {
 	return &orasCmd{